add connection quiz service to user service

+++ quiz-service/src/main/java/com/pikuco/quizservice/dto/QuizResultDto.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public record QuizResultDto(
         List<QuestionResultDto> questions,
-        int participant_id,
+        Long participant_id,
         LocalDateTime passedAt
 ) {
 }
